% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scFM.R
\name{gibbs_scFM}
\alias{gibbs_scFM}
\title{Gibbs sampler for Bayesian segmented Gaussian copula factor model}
\usage{
gibbs_scFM(x, k, m = 1, starting, priors, nmcmc, nburnin)
}
\arguments{
\item{x}{Data matrix (n × p), typically with count-type entries.}

\item{k}{Number of latent factors.}

\item{m}{To be written...}

\item{starting}{A list of starting values for the parameters (e.g., Z, delta, Lambda, U, etc.).}

\item{priors}{A list of prior hyperparameters (e.g., for Sigma and Phi).}

\item{nmcmc}{Number of MCMC iterations.}

\item{nburnin}{Burn-in period.}
}
\value{
A list containing MCMC samples for Lambda and U.
}
\description{
Gibbs sampler for Bayesian segmented Gaussian copula factor model
}
