% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGCFM.R
\name{gibbs_SGCFM}
\alias{gibbs_SGCFM}
\title{Gibbs sampler for Bayesian segmented Gaussian copula factor model}
\arguments{
\item{x}{Data matrix (n × p), typically with count-type entries.}

\item{k}{Number of latent factors.}

\item{starting}{A list of starting values for the parameters (e.g., Z, delta, Lambda, U, etc.).}

\item{priors}{A list of prior hyperparameters (e.g., for Sigma and Phi).}

\item{nmcmc}{Number of MCMC iterations.}
}
\value{
A list containing MCMC samples for Lambda and U.
}
\description{
Gibbs sampler for Bayesian segmented Gaussian copula factor model
}
