\name{scFM-package}
\alias{scFM-package}
\alias{scFM}
\docType{package}
\title{
\packageTitle{scFM}
}
\description{
\packageDescription{scFM}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{scFM}

Maintainer: \packageMaintainer{scFM}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = scFM)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{scFM}}
%\section{Documentation Index}{\packageIndices{scFM}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
